Time: 25 ms (58.63%), Space: 86.1 MB (60.24%) - LeetHub

+++ 3634-minimum-removals-to-balance-array/3634-minimum-removals-to-balance-array.java
@@ -6,7 +6,7 @@ class Solution {
         int left=0;
         int len = 0;
         for(int right=0;right<n;right++){
-            while(nums[right]>k*nums[left]){
+            while((long)nums[right]> (long)k*nums[left]){
                 left++;
             }
             len = Math.max(len,right-left+1);
